requirements.txt
beautifulsoup4==4.12.3
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
soupsieve==2.6
threadpoolctl==3.5.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.20.0
